+++ pjrpc/server/integration/django/urls.py
@@ -1,0 +1,4 @@
+import itertools
+from .sites import jsonrpc_sites
+
+urlpatterns = list(itertools.chain.from_iterable((site.urls for site in jsonrpc_sites)))